PMM-6272 Refactor interfaces to types files
