+++ pmm-app/src/pmm-settings/components/Advanced/Advanced.types.ts
@@ -1,0 +1,8 @@
+import { LoadingCallback } from 'pmm-settings/Settings.service';
+
+export interface AdvancedProps {
+  dataRetention: string;
+  telemetryEnabled: boolean;
+  sttEnabled: boolean;
+  updateSettings: (body: any, callback: LoadingCallback) => void;
+}
