+++ pmm-app/src/pmm-settings/components/Advanced/SwitchRow.types.ts
@@ -1,0 +1,10 @@
+export interface SwitchRowProps {
+  label: string;
+  tooltip: string;
+  tooltipLinkText: string;
+  link: string;
+  checked: boolean;
+  className?: string;
+  dataQa?: string;
+  onChange?: () => void;
+}
