+++ pmm-app/src/pmm-settings/components/AlertManager/AlertManager.types.ts
@@ -1,0 +1,7 @@
+import { LoadingCallback } from 'pmm-settings/Settings.service';
+
+export interface AlertManagerProps {
+  alertManagerUrl: string;
+  alertManagerRules: string;
+  updateSettings: (body: any, callback: LoadingCallback) => void;
+}
