+++ pmm-app/src/pmm-settings/components/MetricsResolution/MetricsResolution.types.ts
@@ -1,0 +1,7 @@
+import { MetricsResolutions } from 'pmm-settings/Settings.types';
+import { LoadingCallback } from 'pmm-settings/Settings.service';
+
+export interface MetricsResolutionProps {
+  metricsResolutions: MetricsResolutions;
+  updateSettings: (body: any, callback: LoadingCallback) => void;
+}
