+++ pmm-app/src/pmm-settings/components/SSHKey/SSHKey.types.ts
@@ -1,0 +1,6 @@
+import { LoadingCallback } from 'pmm-settings/Settings.service';
+
+export interface SSHKeyProps {
+  sshKey: string;
+  updateSettings: (body: any, callback: LoadingCallback) => void;
+}
